interfaz/requirements.txt
Flask
jinja2
gunicorn
Flask-Login
Werkzeug
Pillow
